+++ remix/app/routes/lopu.tsx
@@ -1,0 +1,5 @@
+import { LopuVoiceChat } from '~/components/Lopu/LopuVoiceChat';
+
+export default function Lopu() {
+	return <LopuVoiceChat />;
+}